fix: remove duplicates from symbol table, forbid var redeclarations (#21)
**Description**

This PR modifies the symbol table creation logic by:
- Storing the table as a global variable in the form of a Vec, every
time we parse a symbol we push it to this vector
- If compilation is successful or we exit with an error we first dump
the table into a file
- When parsing the variable declarations inside `init {}` if we find a
symbol that is already in the table we exit with an error.
This PR also modifies the parsing of negative floats and integers:
- The enum variants `NumberNegativeInt` and `NumberNegativeInt` were
removed in favor of using i64 and f32 to store negatives.