refactor: reorganize project structure  (#24)
**Description**

This PR does:

1. Separate the project files into 4 main folders:
- `src/lexer`: holds the files for `rflex` with the lexer implementation
- `src/grammar`: holds the files for `rustemo` with the logic to parse
the program
- `src/compiler`: contains the glue that joins together the lexer and
parser logic, with the global context for the compilation.
    - `src/tests`: contains the files for the integration test

2. Modifies context.rs to hold the compiler context in a struct
CompilerContext
3. Modifies `grammar_actions` to now only hold the functions that are
called when a rule is recognized + adds doc comments to more easily
identify which rule is being parsed
4. Moves all the type declarations from `grammar_actions` to
`src/grammar/types.rs` and adds comments to identify which rules they
represent
5. The parser now uses a rustemo "custom builder" that is a copy of the
default generated builder but the struct that implements the Builder
trait holds the compiler context and in the near future the AST struct.