feat: AST generation for arithmetic expressions (#25)
**Description**

This PR adds the AST generation for arithmetic expressions.
The AST consists of an array of references to nodes, indexed by the enum
AstPtr. It also has two stacks to keep track of expressions and terms.

The 3 main functions to interact with the ast are:
- `assign_node_to_ptr(&mut self, source: AstNodeRef, dest_ptr: AstPtr)`:
**DestPtr = SourcePtr**
- `create_node(&mut self, action: AstAction, left_child_ptr: AstNodeRef,
right_child_ptr: AstNodeRef, dest_ptr: AstPtr)`: **dest_ptr =
createNode(action, left_child, right_child)**
- `create_leaf(&mut self, value: String, dest_ptr: AstPtr)`: **dest_ptr
= createLeaf(value)**